# internal/api/http/openapi.yaml
openapi: 3.1.0
info:
  title: Plowered API
  version: "1.0.0"
  description: |
    Plowered is a self-hosted, AI-native data context platform: catalog,
    orchestration, quality, governance, lineage, ETL/ELT, MCP, all in one.

    All endpoints sit behind the same auth chain — session cookie (web UI)
    or Bearer JWT (integrations). Tenant isolation is enforced server-side
    from the authenticated principal; you never pass tenant_id in payloads
    or query strings.

    Identifiers throughout are UUIDs unless documented otherwise.
  contact:
    name: Plowered
    url: https://github.com/Satyaamm/plowered
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /v1/auth/signup:
    post:
      tags: [auth]
      summary: Create a workspace + first user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "202":
          description: Workspace created; verification email sent.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignupResponse" }
        "400": { description: Validation failed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "409": { description: Email or workspace slug already in use }
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Exchange credentials for a session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Logged in; session cookie set
          headers:
            Set-Cookie: { schema: { type: string } }
        "401": { description: Invalid credentials or unverified email }
...
  /v1/auth/me:
    get:
      tags: [auth]
      summary: Return the authenticated principal
      responses:
        "200":
          description: Principal info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Principal" }
        "401": { description: Not signed in }
  /v1/auth/verify:
    get:
      tags: [auth]
      summary: Consume an email-verification token
      security: []
      parameters:
        - in: query
          name: token
          required: true
          schema: { type: string }
      responses:
        "200": { description: Email verified }
        "400": { description: Invalid or expired token }

  # ---------- catalog ----------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: plowered_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:    { type: string }
        message: { type: string }
    Principal:
      type: object
      properties:
        user_id:   { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        email:     { type: string, format: email }
        full_name: { type: string }
        roles:     { type: array, items: { type: string } }
        groups:    { type: array, items: { type: string } }
        email_verified: { type: boolean }
    SignupRequest:
      type: object
      required: [email, password, workspace_name]
      properties:
        email:
          type: string
          format: email
          maxLength: 256
          description: "Trimmed + lowercased server-side. Must contain @ and a dot."
        password:
          type: string
          minLength: 8
          maxLength: 256
          writeOnly: true
          description: "8+ chars and at least 3 of: lowercase, uppercase, digit, symbol."
        confirm_password:
          type: string
          writeOnly: true
          description: "Optional; if supplied must equal password."
        first_name:
          type: string
          maxLength: 64
          description: "Unicode letters / marks / spaces / apostrophes / hyphens / dots. No digits."
        last_name:
          type: string
          maxLength: 64
          description: "Unicode letters / marks / spaces / apostrophes / hyphens / dots. No digits."
        phone:
          type: string
          pattern: "^[0-9 \\-]{6,20}$"
          description: "Subscriber digits only (6–15 after stripping separators). Optional."
        phone_country:
          type: string
          pattern: "^\\+\\d{1,4}$"
          description: "Dial code, e.g. +1 or +91. Required when phone is set."
        workspace_name:
          type: string
          minLength: 2
          maxLength: 64
        workspace_slug:
          type: string
          description: "Optional; auto-generated from name when blank."
        accept_terms:
          type: boolean
          description: "UI gate; server treats absence as acceptance for legacy clients."
    SignupResponse:
      type: object
      properties:
        tenant_id: { type: string, format: uuid }
        user_id:   { type: string, format: uuid }
        status:    { type: string, enum: [pending_verification] }
        message:   { type: string }
...